AVM Http refactoring WIP
