+++ avm/src/vm/mod.rs
@@ -1,4 +1,5 @@
 pub mod event;
+pub mod http;
 pub mod instruction;
 pub mod memory;
 pub mod opcode;
