+++ avm/src/vm/opcode.rs
@@ -36,43 +36,7 @@ use crate::vm::event::{NOP_ID, BuiltInEvents, EventEmit, HandlerFragment};
 use crate::vm::memory::{FractalMemory, HandlerMemory, CLOSURE_ARG_MEM_START};
 use crate::vm::program::Program;
 use crate::vm::run::EVENT_TX;
-
-#[derive(Debug)]
-pub struct HttpConfig {
-  pub port: u16,
-}
-
-#[derive(Debug)]
-pub struct HttpsConfig {
-  pub port: u16,
-  pub priv_key_b64: String,
-  pub cert_b64: String,
-}
-
-#[derive(Debug)]
-pub enum HttpType {
-  HTTP(HttpConfig),
-  HTTPS(HttpsConfig),
-}
-
-struct HyperAcceptor<'a> {
-  acceptor: Pin<Box<dyn Stream<Item = Result<TlsStream<TcpStream>, io::Error>> + Send + 'a>>,
-}
-
-impl hyper::server::accept::Accept for HyperAcceptor<'_> {
-  type Conn = TlsStream<TcpStream>;
-  type Error = io::Error;
-
-  fn poll_accept(
-    mut self: Pin<&mut Self>,
-    cx: &mut Context,
-    ) -> Poll<Option<Result<Self::Conn, Self::Error>>> {
-    Pin::new(&mut self.acceptor).poll_next(cx)
-  }
-}
-
-static HTTP_CLIENT: Lazy<Client<HttpsConnector<HttpConnector>>> =
-  Lazy::new(|| Client::builder().build::<_, Body>(HttpsConnector::with_native_roots()));
+use crate::vm::http::{HTTP_CLIENT, make_server};
 
 static DS: Lazy<Arc<DashMap<String, Arc<HandlerMemory>>>> =
   Lazy::new(|| Arc::new(DashMap::<String, Arc<HandlerMemory>>::new()));
@@ -3089,69 +3053,7 @@ pub static OPCODES: Lazy<HashMap<i64, ByteOpcode>> = Lazy::new(|| {
   }
   io!(httplsn => fn(_args, hand_mem) {
     Box::pin(async move {
-      match &Program::global().http_config {
-        HttpType::HTTP(http) => {
-          let port_num = http.port;
-          let addr = SocketAddr::from(([0, 0, 0, 0], port_num));
-          let make_svc = make_service_fn(|_conn| async { Ok::<_, Infallible>(service_fn(http_listener)) });
-
-          let bind = Server::try_bind(&addr);
-          match bind {
-            Ok(server) => {
-              let server = server.serve(make_svc);
-              tokio::spawn(async move { server.await });
-              println!("HTTP server listening on port {}", port_num);
-            },
-            Err(ee) => eprintln!("HTTP server failed to listen on port {}: {}", port_num, ee),
-          }
-        },
-        HttpType::HTTPS(https) => {
-          let port_num = https.port;
-          let addr = SocketAddr::from(([0, 0, 0, 0], port_num));
-          let tls_cfg = {
-            let certs = pemfile::certs(
-              &mut BufReader::new(base64::decode(https.cert_b64.as_str()).unwrap().as_slice())
-            );
-            let certs = certs.expect("Failed to load certificate");
-            let key = {
-              let keys = pemfile::pkcs8_private_keys(
-                &mut BufReader::new(
-                  base64::decode(https.priv_key_b64.as_str()).unwrap().as_slice()
-                )
-              );
-              let keys = keys.expect("Failed to load private key");
-              if keys.len() != 1 {
-                panic!("Expected a single private key");
-              }
-              keys[0].clone()
-            };
-            let mut cfg = rustls::ServerConfig::new(rustls::NoClientAuth::new());
-            cfg.set_single_cert(certs, key).unwrap();
-            cfg.set_protocols(&[b"h2".to_vec(), b"http/1.1".to_vec()]);
-            Arc::new(cfg)
-          };
-          let tcp = TcpListener::bind(&addr).await;
-          let tcp = tcp.unwrap();
-          let tls_acceptor = TlsAcceptor::from(tls_cfg);
-          let incoming_tls_stream = stream! {
-            loop {
-              let accept = tcp.accept().await;
-              if accept.is_err() { continue; }
-              let (socket, _) = accept.unwrap();
-              let strm = tls_acceptor.accept(socket).into_failable();
-              let strm_val = strm.await;
-              if strm_val.is_err() { continue; }
-              yield Ok(strm_val.unwrap());
-            }
-          };
-          let make_svc = make_service_fn(|_conn| async { Ok::<_, Infallible>(service_fn(http_listener)) });
-          let server = Server::builder(HyperAcceptor {
-            acceptor: Box::pin(incoming_tls_stream),
-          }).serve(make_svc);
-          println!("HTTPS server listening on port {}", port_num);
-          tokio::spawn(async move { server.await });
-        },
-      }
+      make_server(&Program::global().http_config, http_listener);
       return hand_mem;
     })
   });
